third_party/WebKit/LayoutTests/fast/dom/shadow/hover-active-drag-distributed-nodes-expected.txt

Top-level text Nested text
Other div
PASS backgroundColor is "rgb(0, 128, 0)"
PASS backgroundColor is "rgb(0, 255, 255)"
PASS backgroundColor is "rgb(0, 128, 0)"
PASS backgroundColor is "rgb(255, 0, 0)"
PASS backgroundColor is "rgb(0, 128, 0)"
FAIL textColor should be rgb(0, 128, 0). Was rgb(0, 0, 0).
PASS backgroundColor is "rgb(0, 255, 255)"
FAIL textColor should be rgb(0, 255, 255). Was rgb(0, 0, 0).
PASS backgroundColor is "rgb(0, 128, 0)"
FAIL textColor should be rgb(0, 128, 0). Was rgb(0, 0, 0).
PASS backgroundColor is "rgb(255, 0, 0)"
PASS textColor is "rgb(0, 0, 0)"
PASS backgroundColor is "rgb(0, 128, 0)"
PASS backgroundColor is "rgb(0, 255, 255)"
PASS backgroundColor is "rgb(0, 128, 0)"
PASS backgroundColor is "rgb(255, 0, 0)"
